requirements.txt
pandas>=1.0.0
requests>=2.20.0
python-dotenv>=0.19.0
google-auth>=2.0.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0